requirements.txt
openpyxl==3.1.2
customtkinter==5.2.1
selenium==4.16.0
pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0
packaging==23.2
bcrypt==4.1.2